Updated User and Logging In
Updated the User to include a name field: This way, users can sign in
with either their username or email. Also, this way, users are able to
see other users by their aliases.

Also, when the users signs in, the Header reflects that change. The
sign in button changes to an Account dropdown menu allowing the user to
1) check their profile page, 2) edit their profile page, and 3) sign
out. The user remains logged on until explicitly signing out.

Finally, I added a simple debugging view at the bottom of the website
window to assist debugging while browsing the website.