Added a profile picture to users.
Also checks for the existence of the profile picture. There were simple
tests that did not work after I changed the profile page that just
simply checked for the header1 with the name and email in it. I didn't
think they're relevant anymore, but all I did was comment them out.